+++ Queue.h
@@ -1,6 +1,3 @@
-#include <cstdio>
-using namespace std;
-
 #define SIZE 100
 
 class Queue{
@@ -14,34 +11,3 @@ class Queue{
 		bool isFull();
 		void clear();
 };
-
-bool Queue::append(int x){
-	if(cnt<SIZE){
-		q[tail]=x;
-		tail=(tail+1)%(SIZE+1);
-		++cnt;
-		return 1;
-	}
-	return 0;
-}
-
-int Queue::pop(){
-	int Prehead=head;
-	head=(head+1)%(SIZE+1);
-	--cnt;
-	return q[Prehead];
-}
-
-bool Queue::isEmpty(){
-	return cnt==0;
-}
-
-bool Queue::isFull(){
-	return cnt==SIZE;
-}
-
-void Queue::clear(){
-	while(cnt){
-		pop();
-	}
-}
